requirements.txt
# Wissenschaftliche Bibliotheken
numpy==2.2.3
matplotlib==3.10.1
scipy==1.15.2
sympy==1.13.3

# Datenverarbeitung
pandas==2.2.3
jupyter==1.1.1
notebook==7.3.2

# Visualisierung
seaborn==0.13.2
plotly==6.0.0
ipywidgets==8.1.5
imageio==2.31.5

# Bildverarbeitung
Pillow==11.1.0

# Webentwicklung
flask==3.1.0
requests==2.32.3

# Fortschrittsanzeige
tqdm==4.67.1

# Entwicklungswerkzeuge
pytest==8.3.4
black==25.1.0
flake8==7.1.2
pylint==3.3.4
mypy==1.15.0 